Install `pyright` for type checking and linting
Pyright is the backend that Pylance uses in Visual Studio Code for
type checking and linting. This is beneficial since it means that Visual
Studio Code's Pylance and the CLI tool `pyright` are likely to agree
with each other.

However, since it's implemented in TypeScript instead of Python, it
introduces yet another runtime dependency on `node`, so it might not be
ideal long-term.